app/requirements.txt
langgraph
langchain-core
langchain-openai
pydantic
typing_extensions
